Add server-side toggle for portal prediction
'g_portalPredict' server cvar controls whether predicted portal teleports are enabled. This can be voted by clients. 'vote_allow_portalPredict' can be used to turn off the ability to vote for it.

In-game vote menu was re-designed a bit as it was getting cluttered with vote options, there's a bit more space now for things.