Prevent portals from overlapping if portalteam is set
Don't let players overlap portals if 'portalteam' worldspawn key is set. This solves ambiguity on which portal gets activated when two different players place a portal on top of each other, and 'portalteam' is enabled so that players can use other player's portals.

If 'portalteam' is set to 1, player's portals are cleared upon joining a fireteam, to prevent potentially overlapping portals if they were created before joining a fireteam.