Add dumpEntities command
Dumps all entities in the current map to a file. Takes an optional filename as an argument, defaults to '<mapname>.ent'. 'developer 1' must be set for this to work, and the map must be reloaded, because the entity data as it's presented in the BSP must be stored during init.

This basically acts the same as q3map2's '-exportents' switch, but is much more convenient to use as you can just dump the entities by simply loading a map in game.