Add animation to portal spawning
Portals now spawn with a 200ms long quadratic ease-out animation. This is only for drawing purposes, the actual portal bbox is instantly spawned at maximum size.

This won't work for old demos, as we don't have the timestamp when the portal was spawned. We can't use the client-sided spawntime (e.g. when the entity appeared in a snapshot) because it can cause animations to draw incorrectly in some scenarios, like when teleporting to a new area which has portals already spawned.