Keep spectator list sorted, optimize updating
The spectator list now keeps it's sorting when a clients state changes.

* If a client starts spectating you, they will be placed at the bottom of the active clients list
* If a client stops spectating you, they will be removed from their current place, and the list shifts to fill the gap
* If a player goes inactive, they will be removed from active list, and placed at the bottom of inactive list
* If a player becomes active, they will be removed from the inactive list, and placed at the bottom of active list

The list data is now only updated on 'CS_PLAYERS' configstrings changes, as well as score updates, as these are the only scenarios where the data on the list can change. This reduces the building & sorting of the lists to mostly 3s intervals, rather than it happening every frame.