Fix client inactivity persistence breaking with level time reset
Client inactivity status persistence relied on the last activity timestamp, which broke upon map change if the server was using 'sv_level/serverTimeReset'. We don't actually need to store the timestamp at all, we can simply store the state of client, as the timestamp is only used to determine whether the client should be set to inactive.