Fix client inactivity timer with level time reset, take two
We *DO* need to store the activity timestamp to carry over the idle time from previous level. The previous fix only worked if the client was already inactive. Now if level time reset is enabled, we store a potentially negative timestamp as the last client activity time, which correctly transfers the time the client was idle before a map change onto the next map.