Rework timerun high ping interrupt
'ps.ping' is a completely made up number that has no actual relevance to the client's connection on the current frame, therefore we can't use it to reliably determine the threshold on whether the client is lagging or not. Instead, check the 'usercmd->serverTime - ps.commandTime' delta, and determine the amount of lag from that. If we have over 600ms of delta, interrupt the run.