Fix infinite loop in changelog formatting
If there was a line which started with a word that would exceed the max width of a line, the parser would get stuck on infinite loop, as it was constantly trying to split the line from the previous whitespace, without ever consuming the line further. In such cases, just split at the line end and break a word in the middle.

This likely happened as part of c54d3fcb3702d60377c58821ec4db351e4e1d35c, since I do remember testing every single changelog while implementing this, but the size decrease caused the parser to fail in some scenarios, like in 3.1.0 and 2.3.0 changelogs.