Fix client inactivity drop position restore not working
This code has been in the mod since 2013 (2.0.3 RC1), but it has never worked because the position restore was made in an incorrect location. Clients are now correctly placed back to the position they were moved to spectators from due to inactivity. This persist for one spawn for a team (you're allowed to join wrong team initially, it won't invalidate the position - only joining the team you were at the time of drop uses up the position and invalidates it) and takes precedence over a position from 'etj_autoLoad'. In addition to the original implementation, it also now saves and restores your viewangles.