Fix server crash if target_spawn_relay entity was removed
The pointers which pointed to the spawn relay entities were not invalidated when a 'target_spawn_relay' entity was removed, which caused a crash as it was now a null pointer.

This now handled for 'delete', 'kill' and 'remove' script actions, as well as 'target_kill'. I can't think of any other ways to delete entities from the top of my head.