Fix spawning projectiles on the same frame as timerun starts
This prevents a frame-perfect exploit with timerun start, which allowed players to spawn projectiles on the same frame as they started a timerun.

Because the server processes client thinks before it runs a game frame, it was possible for players to spawn a projectile on the same exact frame as they started a timerun. These projectiles would not get deleted by the standard projectile deletion on timerun start, as they would be spawned in later on that same frame. This allowed player to for example boost themselves with panzerfaust explosions, effectively bypassing any start speed restrictions.