Add support for up to 511 unique shaders for remapping
Adds support for up to 511 unique shaders that the game can reserve to be remapped. This is done by simply communicating the extra shaders, indices and states via regular server commands. 511 (or 512 to be precise but configstrings never use index 0) is the practical limit for this, as shader index for certain entities which register a shader to use (e.g. 'target_smoke' with custom shader) use 'es->modelindex2' to communicate the shader, which is networked as 9 bits.