Fix re-vote potentially reducing from wrong team's vote count
In scenarios where the client voted while on a team, changed to spectators and re-voted (or vice versa), the yes/no vote count would be deducted from the wrong count. Track the team that the client was last on when they voted, and reduce from correct count. Also fixes re-voting 'no' on RTV not being handled correctly.