Fix buffer overflow with autortv
Autortv sent a 2-byte buffer as 'arg2' to the callvote function, which would cause a buffer overflow because the code that should copy the name of a custom vote list to 'arg2' wasn't checking whether 'arg2' actually contained the list name we wanted to vote for. If 'arg2' is empty, nothing should be copied to it, but for safety, send a 256-byte buffer from auto-rtv too, as the rest of the code path expects it to be that size.