Add custom command menu
This menu can hold up to 40 custom, user defined commands. The idea is that users can store commands here that are "not used frequently enough to warrant a dedicated bind" but "used frequently enough that typing them in console is annoying". The commands are stored in 'custom_commands.dat' file.

Commands:
* 'openCustomCommandMenu' - opens the menu
* 'addCustomCommand <name> <command>' - adds a new command to the menu
* 'deleteCustomCommand <number>' - deletes a custom command by index
* 'listCustomCommands' - lists available commands

The cvar 'etj_ccMenu_rememberPage' can be used to toggle whether the game remembers the last opened page on the menu, or always opens the first page.