Reimplement !admintest, !finger & !listusernames
Some of the simpler commands to integrate to the new system for starters. Not a big fan of 'SessionV2' handling '!listusernames', I don't think it really belongs there, but the Commands class would need it's own repository handler to keep it contained there, which is out of scope for this.