Reimplement !finduser
Reworded the output slightly too to indicate that the results contain max 20 matches. Also fixes a bug with the old implementation where the queried name wasn't sanitized, which would always result in no results on colored names, because db query was done against 'clean_name' column. You could argue that this was working "as expected", but it's a bit unintuitive and frankly stupid.